chore: Update CodeRepositoryPage form fields
This commit updates the CodeRepositoryPage form fields by removing the "Main Branch Name" field and adding a "Select Service" dropdown with a description. The "Main Branch Name" field is no longer necessary and the "Select Service" dropdown provides a better user experience by allowing users to choose a service from the Service Catalog. These changes improve the usability and clarity of the form.