refactor: Add support for reportDataInDays in StatusPage model
This commit adds the `reportDataInDays` property to the `StatusPage` model in the `StatusPage.ts` file. The `reportDataInDays` property allows users to specify the number of days of data to include in the report. This change enhances the reporting functionality of the application by providing more flexibility in generating reports for status pages.