refactor: Add Reviews module and update home page to display reviews
This commit adds a new Reviews module to the Home feature set. The Reviews module contains a list of reviews that will be displayed on the home page. The home page is updated to include the Reviews module and pass the reviews list to the view template. This enhancement improves the user experience by showcasing positive feedback from customers.

Files modified:
- App/FeatureSet/Home/Index.ts
- App/FeatureSet/Home/Views/Partials/proudly-open-source.ejs
- App/FeatureSet/Home/Views/Partials/review.ejs
- App/FeatureSet/Home/Utils/Reviews.ts