refactor: Update UptimeUtil to round uptime percentage to specified precision
The UptimeUtil class in the Common/Utils/Uptime directory has been updated to include a new method called roundToPrecision. This method takes a number and a precision value and rounds the number to the specified precision. This method is used in the calculateUptimePercentage method to round the uptime percentage to the precision specified in the data.group object.

This change ensures that the uptime percentage is displayed with the correct precision, providing more accurate information to users.

Refactor the UptimeUtil class to include the roundToPrecision method and update the calculateUptimePercentage method to use it.

Fixes #123